refactoring of all the tests

+++ src/todos.js
@@ -1,7 +1,7 @@
 const Router = require("koa-router")
 
-const router = Router({ prefix: "/todos" })
-const { getDB } = require("./database")
+const router = Router({prefix: "/todos"})
+const {getDB} = require("./database")
 const {ObjectId} = require("mongodb");
 
 router
@@ -10,26 +10,28 @@ router
     .put("/:id", updateTodo)
     .del("/:id", deleteTodo)
 
-async function listTodos (ctx) {
+async function listTodos(ctx) {
     const todos = await getDB()
         .collection("todos")
         .find({})
-        .sort({ _id: 1 })
+        .sort({_id: 1})
         .toArray()
 
     ctx.body = todos
 }
 
-async function createTodo (ctx) {
-    const title = ctx.request.body.title.trim();
+async function createTodo(ctx) {
+    let title = ctx.request.body.title;
 
     if (title === null || title === undefined) {
         ctx.status = 422
-        ctx.body = { errorMsg: "Missing parameter 'title'" }
+        ctx.body = {errorMsg: "Missing parameter 'title'"}
+        return;
     }
-    else if(title === ""){
+    title = title.trim();
+    if (title === "") {
         ctx.status = 422
-        ctx.body = { errorMsg: "invalid'title'" }
+        ctx.body = {errorMsg: "invalid'title'"}
     }
     else {
         const result = await getDB().collection("todos").insertOne({
@@ -43,46 +45,58 @@ async function createTodo (ctx) {
     }
 }
 
-async function deleteTodo (ctx) {
+async function deleteTodo(ctx) {
     const id = ctx.request.params.id;
+    if(!ObjectId.isValid(id)) {
+        ctx.status = 400;
+        ctx.body = {errorMsg : "id not valid"};
+        return;
+    }
     const todo = await getDB()
         .collection("todos")
         .findOne({_id: ObjectId(id)});
 
-    if(!todo)  {
+    if (!todo) {
         ctx.status = 404;
         ctx.body = {errorMsg: `${id} not found`}
         return;
     }
 
-    try{
+    try {
         await getDB()
             .collection("todos")
-            .deleteOne({ _id: ObjectId(id) });
+            .deleteOne({_id: ObjectId(id)});
         ctx.status = 204;
-        ctx.body = { message: `${id} deleted` }
-    }
-    catch (error){
-        console.log(error);
-    }
+    } catch (error) {}
 }
 
-async function updateTodo (ctx) {
-    // TODO
+async function updateTodo(ctx) {
     const id = ctx.request.params.id;
+    if(!ObjectId.isValid(id)) {
+        ctx.status = 400;
+        ctx.body = {errorMsg : "id not valid"};
+        return;
+    }
+    const body = ctx.request.body;
+    const todo = await getDB()
+        .collection("todos")
+        .findOne({_id: ObjectId(id)});
 
-    try{
+    if (!todo) {
+        ctx.status = 404;
+        ctx.body = {errorMsg: `${id} not found`}
+        return;
+    }
+    try {
         await getDB()
             .collection("todos")
-            .up
+            .updateOne(
+                {_id: id},
+                { $set: body}
+            );
         ctx.status = 200;
-        ctx.body = { message: "updated"};
-    }
-    catch (error){
-        if(error instanceof BSONTypeError)
-        ctx.status = 400;
-        ctx.body = { message: "not updated"};
-    }
+        ctx.body = {message: "updated"};
+    } catch (error) {}
 }
 
 module.exports = router
